Recuperar contraseña al peluche

+++ app/controllers/SessionController.php
@@ -58,10 +58,8 @@ class SessionController extends \Phalcon\Mvc\Controller
 			));
 			
 			if ($user) {
-				$this->session->set('userid', $user->idUser);
-				
 				$cod = uniqid();
-				$url = $this->url->get('session/edituseraccount');
+				$url = $this->url->get('session/validaterequest');
 				$url .= '/' . $cod;
 				
 				$recoverObj = new Tmprecoverpass();
@@ -80,19 +78,40 @@ class SessionController extends \Phalcon\Mvc\Controller
 				else {
 					$message = new AdministrativeMessages();
 					$message->createRecoverpassMessage($url, $user->email);
-					$message->sendMessage();
+//					$message->sendMessage();
 				}
-				$this->flashSession->success('Se le ha enviado un correo electronico con instrucciones');
+				$this->flashSession->success('Se ha enviado un correo electronico con instrucciones para recuperar la contraseña');
 			}
 		}
 	}
 	
-	public function edituseraccountAction($uniq = null)
+	public function validaterequestAction($unique)
 	{
-		$this->logger->log('Uniq' . $uniq);
+		$this->logger->log('0');
+		$url = Tmprecoverpass::findFirst(array(
+			'conditions' => 'idTmpRecoverPass = ?1',
+			'bind' => array(1 => $unique)
+		));
 		
-		$this->view->setVar('uniq', $uniq);
+		$time = strtotime("-30 minutes");
+		$this->logger->log('1');
+		if ($url && ($url->date <= $time || $url->date >= $time)) {
+			$this->logger->log('2');
+			$this->session->set('idUser', $url->idUser); 
+			$this->view->setVar('uniq', $unique);
+		}
+		else {
+			$this->logger->log('3');
+			return $this->response->redirect('error');
+		}
+	}
+	
+	public function setnewpassAction()
+	{
+		$this->logger->log('0');
 		if ($this->request->isPost()) {
+			$uniq = $this->request->getPost("uniq");
+			$this->logger->log('1');
 			$url = Tmprecoverpass::findFirst(array(
 				'conditions' => 'idTmpRecoverPass = ?1',
 				'bind' => array(1 => $uniq)
@@ -101,45 +120,76 @@ class SessionController extends \Phalcon\Mvc\Controller
 			$time = strtotime("-30 minutes");
 			
 			if ($url && ($url->date <= $time || $url->date >= $time)) {
+				$this->logger->log('2');
 				$pass = $this->request->getPost("pass");
 				$pass2 = $this->request->getPost("pass2");
 
 				if (empty($pass)||empty($pass2)){
+					$this->logger->log('3');
 					$this->flashSession->error("Ha enviado campos vacíos, por favor verifique la información");
+					$this->dispatcher->forward(array(
+						"controller" => "session",
+						"action" => "validaterequest",
+						"params" => array($uniq)
+					));
 				}
 				else if (strlen($pass) < 8 || strlen($pass) > 40) {
+					$this->logger->log('4');
 					$this->flashSession->error("La contraseña es muy corta o muy larga, esta debe tener mínimo 8 y máximo 40 caracteres, por favor verifique la información");
+					$this->dispatcher->forward(array(
+						"controller" => "session",
+						"action" => "validaterequest",
+						"params" => array($uniq)
+					));
 				}	
 				else if ($pass !== $pass2) {
+					$this->logger->log('5');
 					$this->flashSession->error("Las contraseñas no coinciden, por favor verifique la información");
+					$this->dispatcher->forward(array(
+						"controller" => "session",
+						"action" => "validaterequest",
+						"params" => array($uniq)
+					));
 				}
 				else {
-					$idUser = $this->session->get('userid');
-
+					$this->logger->log('6');
+					
+					$idUser = $this->session->get('idUser');
+					
 					$user = User::findFirst(array(
 						'conditions' => 'idUser = ?1',
 						'bind' => array(1 => $idUser)
 					));
+					
+					if ($user) {
+						$this->logger->log('Hash: ' . $this->security2->hash($pass));
+						$this->logger->log("There's user");
+						$user->password = $this->security2->hash($pass);
 
-					$user->password = $this->security2->hash($pass);
+						if (!$user->save()) {
+							$this->logger->log('7');
+							$this->flashSession->notice('Ha ocurrido un error, contacte con el administrador');
+							foreach ($user->getMessages() as $msg) {
+								$this->logger->log('Error while recovering user password' . $msg);
+							}
+						}
+						else {
+							$this->logger->log('8');
 
-					if (!$user->save()) {
-						$this->flashSession->notice('Ha ocurrido un error, contacte con el administrador');
-						foreach ($user->getMessages() as $msg) {
-							$this->logger->log('Error while recovering user password' . $msg);
+							$url->delete();
+							$idUser = $this->session->remove('idUser');
+							$this->flashSession->notice('Se ha actualizado el usuario exitosamente');
+							return $this->response->redirect('index');
 						}
 					}
 					else {
-						$idUser = $this->session->remove('userid');
-						
-						$url->delete();
-						
-						$this->flashSession->notice('Se ha actualizado el usuario exitosamente');
-						return $this->response->redirect('index');
+						$this->logger->log('9');
+						return $this->response->redirect('error');
 					}
 				}
 			}
 			else {
+				$this->logger->log('10');
 				return $this->response->redirect('error');
 			}
 		}
